+++ Module 5 Assignment/src/FibRecursion.java
@@ -6,14 +6,14 @@
  */
 public class FibRecursion {
 	//Declare variable
-	int fib;
+	long fib;
 	/**
 	 * Method to find Fibonacci number using recursion approch
 	 * 
 	 * @param fib fibonacci number at a given nth nubmer
 	 * @return fib
 	 */
-	public int getFibNumber(int n)
+	public long getFibNumber(int n)
 	{
 		if (n == 1) return 0;
 		if (n == 2) return 1;
